feat(kani): add proofs for verify_add_used and notifications
The virtIO queue has two rings - Avail and Used.
Avail is written by the driver and read by the device.
* It has a `used_event` field that the driver writes to tell the device how many buffers to process before notifying it
Used is written by the device and read by the driver.
* It has a `avail_event` field that the device writes to tell the driver how many buffers it can push for processing before notifying the device.

The `enable_notification` method is called by the device and tells the driver to send a notification when the `avail_ring.next_avail` slot is filled up by the driver.

The `needs_notification` method is called by the driver and tell the device to send a notification if the condition c for triggering a call is true.

```
 c = old =< used_event \< next_used

where,
old - is the previous entry on which the driver was notified by device(`needs_notification` was called).
used_event - is the driver written index for which it wants to be notified
next_used - is the next entry that will be put in the used ring
```

Note that the ordering logic has to account for wrapping. See
`pred_in_wrapping_range` for details.

Signed-off-by: Siddharth Priya <s2priya@uwaterloo.ca>